# [performance] Add allocation reporting