--- a/multibody/benchmarking/cassie.cc
+++ b/multibody/benchmarking/cassie.cc
@@ -7,6 +7,7 @@
 #include "drake/multibody/parsing/parser.h"
 #include "drake/multibody/plant/multibody_plant.h"
 #include "drake/tools/performance/fixture_common.h"
+#include "drake/tools/performance/fixture_memory.h"
 
 namespace drake {
 namespace multibody {
@@ -50,6 +51,7 @@ class Cassie : public benchmark::Fixture {
   void SetUp(BenchmarkStateRef state) override {
     SetUpNonZeroState();
     SetUpGradientsOrVariables(state);
+    tools::performance::TareMemoryManager();
   }
 
  protected: