--- a/tools/performance/fixture_memory_no.cc
+++ b/tools/performance/fixture_memory_no.cc
@@ -1,0 +1,13 @@
+#include "drake/tools/performance/fixture_memory.h"
+
+namespace drake {
+namespace tools {
+namespace performance {
+
+void EnableMemoryManager() {}
+
+void TareMemoryManager() {}
+
+}  // namespace performance
+}  // namespace tools
+}  // namespace drake