--- a/tools/workspace/googlebenchmark/patches/console_allocs.patch
+++ b/tools/workspace/googlebenchmark/patches/console_allocs.patch
@@ -1,0 +1,30 @@
+If a memory manager is installed then shows allocs in the summary table.
+
+--- src/console_reporter.cc
++++ src/console_reporter.cc
+@@ -53,9 +53,11 @@ bool ConsoleReporter::ReportContext(const Context& context) {
+ }
+ 
+ void ConsoleReporter::PrintHeader(const Run& run) {
++  const bool show_allocs = (run.memory_result != nullptr);
+   std::string str =
+-      FormatString("%-*s %13s %15s %12s", static_cast<int>(name_field_width_),
+-                   "Benchmark", "Time", "CPU", "Iterations");
++      FormatString("%-*s %13s %15s %*s%12s", static_cast<int>(name_field_width_),
++                   "Benchmark", "Time", "CPU", show_allocs ? 10 : 0,
++                   show_allocs ? "Allocs " : "", "Iterations");
+   if (!run.counters.empty()) {
+     if (output_options_ & OO_Tabular) {
+       for (auto const& c : run.counters) {
+@@ -163,6 +165,11 @@ void ConsoleReporter::PrintRunData(const Run& result) {
+   }
+ 
+   if (!result.report_big_o && !result.report_rms) {
++    const bool show_allocs = (result.memory_result != nullptr);
++    if (show_allocs) {
++      const std::string s = HumanReadableNumber(result.allocs_per_iter, Counter::kIs1000);
++      printer(Out, COLOR_YELLOW, "%7s   ", s.c_str());
++    }
+     printer(Out, COLOR_CYAN, "%10lld", result.iterations);
+   }
+