Follow-up from #17487 to restore heap reporting.

Looking carefully at the `compare.py` report for the `CassieAutoDiff/MassMatrix/1` data with repetitions=100, it definitively shows that this new instrumentation costs exactly 3 extra CPU cycles per allocation.  On my 3 GHz machine, that's +1 ns per allocation, scaled up by 31,400 allocations is a difference of +31us on benchmark (~2%).

This probably the most extreme case; of all the benchmarks, MassMatrix has one of the highest fraction of allocations per instruction.

We could imagine automatically offering benchmark binaries that do not link in the malloc hooks, but for me the 3 cycle penalty is a reasonable enough one to leave unconditionally enabled.  Any relative before/after measurements for performance improvements should still hold true.

---

Sample output for cassie now:
```
----------------------------------------------------------------------------------------
Benchmark                                    Time             CPU    Allocs   Iterations
----------------------------------------------------------------------------------------
CassieDouble/MassMatrix/0                 12.8 us         12.8 us         0        49308
CassieDouble/InverseDynamics/0            16.5 us         16.5 us         3        42641
CassieDouble/ForwardDynamics/0            35.9 us         35.9 us         5        19442
CassieAutoDiff/MassMatrix/0                402 us          402 us         0         1753
CassieAutoDiff/MassMatrix/1               1256 us         1255 us     31.4k          530
CassieAutoDiff/MassMatrix/2                410 us          410 us         0         1707
CassieAutoDiff/MassMatrix/3               1491 us         1491 us     31.4k          467
CassieAutoDiff/InverseDynamics/0           504 us          503 us         3         1358
CassieAutoDiff/InverseDynamics/1          1492 us         1491 us     36.9k          475
CassieAutoDiff/InverseDynamics/2           858 us          858 us     10.8k          817
CassieAutoDiff/InverseDynamics/3          1897 us         1896 us     37.4k          377
CassieAutoDiff/InverseDynamics/4           720 us          719 us      6.9k          981
CassieAutoDiff/InverseDynamics/5          1856 us         1855 us     37.5k          376
CassieAutoDiff/InverseDynamics/6          1011 us         1010 us     12.0k          686
CassieAutoDiff/InverseDynamics/7          2287 us         2285 us     38.0k          309
CassieAutoDiff/ForwardDynamics/0           821 us          820 us        73          856
CassieAutoDiff/ForwardDynamics/1          2563 us         2562 us     55.9k          274
CassieAutoDiff/ForwardDynamics/2          1209 us         1209 us     11.6k          581
CassieAutoDiff/ForwardDynamics/3          3278 us         3276 us     57.4k          215
CassieAutoDiff/ForwardDynamics/8           897 us          897 us      3.4k          775
CassieAutoDiff/ForwardDynamics/9          2964 us         2962 us     56.2k          238
CassieAutoDiff/ForwardDynamics/10         1225 us         1224 us     11.6k          567
CassieAutoDiff/ForwardDynamics/11         3524 us         3521 us     57.7k          198
CassieExpression/MassMatrix/0             1772 us         1771 us      7.0k          397
CassieExpression/MassMatrix/1             8557 us         8552 us     48.2k           83
CassieExpression/MassMatrix/2             1769 us         1768 us      7.0k          399
CassieExpression/MassMatrix/3             8468 us         8465 us     48.2k           83
CassieExpression/InverseDynamics/0        2275 us         2274 us      8.4k          308
CassieExpression/InverseDynamics/1       29690 us        29675 us     73.4k           24
CassieExpression/InverseDynamics/2        7509 us         7502 us     48.6k           91
CassieExpression/InverseDynamics/3       27136 us        27115 us     74.5k           26
CassieExpression/InverseDynamics/4        4294 us         4291 us     32.5k          162
CassieExpression/InverseDynamics/5       27763 us        27745 us     74.2k           26
CassieExpression/InverseDynamics/6       10374 us        10368 us     50.3k           64
CassieExpression/InverseDynamics/7       26030 us        26011 us     75.2k           28
CassieExpression/ForwardDynamics/0        3625 us         3623 us     13.8k          189
CassieExpression/ForwardDynamics/2       17187 us        17175 us     88.3k           41
CassieExpression/ForwardDynamics/8        4738 us         4736 us     29.6k          147
CassieExpression/ForwardDynamics/10      16944 us        16932 us     88.4k           42
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/17518)
<!-- Reviewable:end -->