@jwnimmer-tri 2022-07-07
+@rpoyner-tri what do you think about this?

Another tack would be to use [`RegisterMemoryManager`](https://github.com/google/benchmark/blob/main/docs/user_guide.md#memory-usage) to report it that way.  I don't know if you tried exploring that API yet?