@jwnimmer-tri 2022-07-07
Yup, I had seen that in the history, but before the change I did run the benchmark a bunch of times with and without the counters it didn't appear to affect the reported timings.  Maybe I didn't run it enough, or made a PEBKAC?

I would like to get a bit more heap telemetry than just the two or three samples we had before -- I think some cases will have a bit more variation as things settle out.

Maybe I'll take a shot at a `RegisterMemoryManager` that doesn't use `LimitMalloc` at all, to be more obviously on the side of "will never affect the timings in any meaningful way".

Aside: I'm also not sure why I used a mutex in `limit_malloc`.  Upon re-reading it, I think a simple `std::atomic<std::shared_ptr>` (spelled in the awkward C++17 form for now) would suffice.