@jwnimmer-tri 2022-07-14
Once the #17581 refactor is merged, I'll return to here and try to (1) measure the instrumentation overhead, and (2) add instructions to display the heap results.