@jwnimmer-tri 2023-02-22
I just pushed a rebase, plus some clean-ups.  I think it's ready for high-level inspection now @rpoyner-tri.

The basic heap reporting is available the default output table.  If you look into to the json report, there's additional heap data as well (beyond just the number of allocs).  I think maybe the current reporting is good enough.

I haven't tried to measure the additional instrumentation overhead.  Possibly you'd be game for trying to do that?