@jwnimmer-tri 2023-02-22
Yup.  With `limit_malloc` the overhead would have been more.  I'm hoping that an extra couple of integer operations + predicted function call will be way down in the noise.