@jwnimmer-tri 2023-04-04
> > I think it's ready for high-level inspection now @rpoyner-tri. I haven't tried to measure the additional instrumentation overhead for leaving this unconditionally enabled. Possibly you'd be game for trying to do that?
>
> I'll see what I can recall or rediscover.

@rpoyner-tri are you still planning to look into this?

Or we could ask @sherm1 to join in and help, since his major goals these days are runtime performance.