@jwnimmer-tri 2023-04-04
> Am I correct that this has zero overhead if we don't enable tracking? If so I like it!

See above.  I estimated the penalty as 3 cycles per malloc.  I was hoping Rico (now you) could double-check that, and then also bless that we don't care.