@jwnimmer-tri 2023-04-04
> Is that because it overrides malloc(), etc. unconditionally even when we're not benchmarking

Yes.  As of this PR, in benchmarking programs (not application code) we unconditionally redirect malloc.  So all benchmarks would be tallying the mallocs.  The command-line flag here only toggles some one-time begin/end stats and the output reporting.

> I wouldn't object to burning a few cycles per malloc during benchmarking but it seems a shame to do that otherwise. Is it possible to override only when benchmarking or testing?

I think maybe you thought that this hooks malloc in production code?  The hook is _only_ in the benchmarking programs.