@jwnimmer-tri 2023-06-02
@drake-jenkins-bot retest this please

+@ggould-tri for platform review or delegation per schedule, please.